+++ src/components/data/data-utils.js
@@ -18,8 +18,7 @@ import { Sortable, TextCell } from 'src/components/table'
 import TooltipTrigger from 'src/components/TooltipTrigger'
 import { UriViewerLink } from 'src/components/UriViewer'
 import ReferenceData from 'src/data/reference-data'
-import { Ajax } from 'src/libs/ajax'
-import { canUseWorkspaceProject } from 'src/libs/ajax/ajax-common'
+import { Ajax, canUseWorkspaceProject } from 'src/libs/ajax'
 import colors from 'src/libs/colors'
 import { reportError } from 'src/libs/error'
 import Events from 'src/libs/events'
